pyproject.toml
[project]
name = "home-scrape"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "rtoml>=0.13.0",
]
